ajout de gold à la mort des ennemis

+++ Projet_DJV2/Assets/Scripts/ScriptableObject/EnemyData.cs
@@ -6,6 +6,6 @@ public class EnemyData : ScriptableObject
     public int score;
     public int maxHealth;
     public float speed;
-    public float reward;
+    public int reward;
 }
